[IMP] sale_automatic_workflow: - Move register_payment option from sale_automatic_workflow_payment_mode to sale_automatic_workflow. - Add _register_payments function in sale_automatic_workflow. TT31208
