+++ sale_automatic_workflow_payment_mode/__manifest__.py
@@ -12,7 +12,6 @@
     "data": [
         "data/automatic_workflow_data.xml",
         "views/account_payment_mode_views.xml",
-        "views/sale_workflow_process_view.xml",
     ],
     "installable": True,
     "auto_install": True,
