+++ sale_automatic_workflow_payment_mode/models/automatic_workflow_job.py
@@ -3,10 +3,7 @@
 
 import logging
 
-from odoo import api, fields, models
-from odoo.tools.safe_eval import safe_eval
-
-from odoo.addons.sale_automatic_workflow.models.automatic_workflow_job import savepoint
+from odoo import models
 
 _logger = logging.getLogger(__name__)
 
@@ -14,57 +11,21 @@ _logger = logging.getLogger(__name__)
 class AutomaticWorkflowJob(models.Model):
     _inherit = "automatic.workflow.job"
 
-    @api.model
-    def run_with_workflow(self, sale_wkf):
-        workflow_domain = [("workflow_process_id", "=", sale_wkf.id)]
-        res = super(AutomaticWorkflowJob, self).run_with_workflow(sale_wkf)
-        if sale_wkf.register_payment:
-            self._register_payments(
-                safe_eval(sale_wkf.payment_filter_id.domain) + workflow_domain
-            )
-        return res
-
-    @api.model
-    def _register_payments(self, payment_filter):
-        invoice_obj = self.env["account.move"]
-        invoices = invoice_obj.search(payment_filter)
-        _logger.debug("Invoices to Register Payment: %s", invoices.ids)
-        for invoice in invoices:
-            partner_type = (
-                invoice.type in ("out_invoice", "out_refund")
-                and "customer"
-                or "supplier"
-            )
+    def _prepare_dict_account_payment(self, invoice):
+        vals = self._prepare_dict_account_payment(invoice)
+        if invoice.payment_mode_id:
             payment_mode = invoice.payment_mode_id
+            vals["payment_type"] = payment_mode.payment_type
+            vals["payment_method_id"] = payment_mode.payment_method_id.id
+            vals["journal_id"] = payment_mode.fixed_journal_id.id
+        return vals
 
-            if not payment_mode.fixed_journal_id:
-                _logger.debug(
-                    "Unable to Register Payment for invoice %s: "
-                    "Payment mode %s must have fixed journal",
-                    invoice.id,
-                    payment_mode.id,
-                )
-                return
-
-            communication = (
-                invoice.name
-                if invoice.type in ("out_invoice", "out_refund")
-                else invoice.reference
+    def _register_payment_invoice(self, invoice):
+        if not invoice.payment_mode_id.fixed_journal_id:
+            _logger.debug(
+                "Unable to Register Payment for invoice %s: "
+                "Payment mode %s must have fixed journal",
+                invoice.id,
+                invoice.payment_mode_id.id,
             )
-
-            with savepoint(self.env.cr):
-                payment = self.env["account.payment"].create(
-                    {
-                        "invoice_ids": [(6, 0, invoice.ids)],
-                        "amount": invoice.residual,
-                        "payment_date": fields.Date.context_today(self),
-                        "communication": communication,
-                        "partner_id": invoice.partner_id.id,
-                        "partner_type": partner_type,
-                        "payment_type": payment_mode.payment_type,
-                        "payment_method_id": payment_mode.payment_method_id.id,
-                        "journal_id": payment_mode.fixed_journal_id.id,
-                    }
-                )
-                payment.post()
-        return
+        return super()._register_payment_invoice(invoice)
