--- sale_automatic_workflow_payment_mode/views/sale_workflow_process_view.xml
@@ -1,58 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<!--
-sale_automatic_workflow_payment_mode for Odoo
-Copyright (C) 2016 Sodexis <dev@sodexis.com>
-The licence is in the file __openerp__.py -->
-<odoo>
-    <record id="sale_workflow_process_view_form" model="ir.ui.view">
-        <field name="name">sale.order.form.automatic.inherit</field>
-        <field name="model">sale.workflow.process</field>
-        <field
-            name="inherit_id"
-            ref="sale_automatic_workflow.sale_workflow_process_view_form"
-        />
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='sale_done']" position="before">
-                <field name="register_payment" />
-                <label
-                    for="payment_filter_id"
-                    attrs="{'required':[('register_payment','=',True)], 'invisible':[('register_payment','!=',True)]}"
-                />
-                <div
-                    attrs="{'required':[('register_payment','=',True)], 'invisible':[('register_payment','!=',True)]}"
-                >
-                    <field
-                        name="payment_filter_domain"
-                        widget="domain"
-                        attrs="{'invisible': [('payment_filter_id', '=', False)]}"
-                        options="{'model': 'account.move'}"
-                    />
-                    <div class="oe_edit_only oe_inline">
-                        Set selection based on a search filter:
-                        <field
-                            name="payment_filter_id"
-                            domain="[('model_id', '=', 'account.move')]"
-                            class="oe_inline"
-                            context="{'default_model_id': 'account.move', 'default_active': False, 'active_test': False}"
-                            can_create="true"
-                            can_write="true"
-                        />
-                    </div>
-                </div>
-            </xpath>
-        </field>
-    </record>
-    <record id="sale_workflow_process_view_tree" model="ir.ui.view">
-        <field name="name">sale_workflow_process.tree.inherit</field>
-        <field name="model">sale.workflow.process</field>
-        <field
-            name="inherit_id"
-            ref="sale_automatic_workflow.sale_workflow_process_view_tree"
-        />
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='validate_invoice']" position="after">
-                <field name="register_payment" />
-            </xpath>
-        </field>
-    </record>
-</odoo>
